Added bootstrap related functions and tests.
